Centralize file uploads into a single function
