+++ app/apps/server/communication/rest.js
@@ -1,8 +1,8 @@
 import { Meteor } from 'meteor/meteor';
 import { HTTP } from 'meteor/http';
-import Busboy from 'busboy';
 
 import { API } from '../../../api/server';
+import { getUploadFormData } from '../../../api/server/lib/getUploadFormData';
 import { getWorkspaceAccessToken, getUserCloudAccessToken } from '../../../cloud/server';
 import { settings } from '../../../settings';
 import { Info } from '../../../utils';
@@ -24,24 +24,6 @@ export class AppsRestApi {
 		this.loadAPI();
 	}
 
-	_handleMultipartFormData(request) {
-		const busboy = new Busboy({ headers: request.headers });
-		return Meteor.wrapAsync((callback) => {
-			const formFields = {};
-			busboy.on('file', Meteor.bindEnvironment((fieldname, file) => {
-				const fileData = [];
-				file.on('data', Meteor.bindEnvironment((data) => {
-					fileData.push(data);
-				}));
-
-				file.on('end', Meteor.bindEnvironment(() => { formFields[fieldname] = Buffer.concat(fileData); }));
-			}));
-			busboy.on('field', (fieldname, val) => { formFields[fieldname] = val; });
-			busboy.on('finish', Meteor.bindEnvironment(() => callback(undefined, formFields)));
-			request.pipe(busboy);
-		})();
-	}
-
 	async loadAPI() {
 		this.api = new API.ApiClass({
 			version: 'apps',
@@ -56,7 +38,6 @@ export class AppsRestApi {
 	addManagementRoutes() {
 		const orchestrator = this._orch;
 		const manager = this._manager;
-		const multipartFormDataHandler = this._handleMultipartFormData;
 
 		const handleError = (message, e) => {
 			// when there is no `response` field in the error, it means the request
@@ -239,7 +220,9 @@ export class AppsRestApi {
 						return API.v1.failure({ error: 'Direct installation of an App is disabled.' });
 					}
 
-					const formData = multipartFormDataHandler(this.request);
+					const formData = Promise.await(getUploadFormData({
+						request: this.request,
+					}));
 					buff = formData?.app;
 					permissionsGranted = (() => {
 						try {
@@ -460,7 +443,9 @@ export class AppsRestApi {
 						return API.v1.failure({ error: 'Direct updating of an App is disabled.' });
 					}
 
-					const formData = multipartFormDataHandler(this.request);
+					const formData = Promise.await(getUploadFormData({
+						request: this.request,
+					}));
 					buff = formData?.app;
 					permissionsGranted = (() => {
 						try {
